Wizzup provided a THREE LINE patch to get it working on x64 after months of bitching... That lazy ass.

+++ src/Color.cpp
@@ -39,8 +39,8 @@ extern int client_height;
     return bmp;
 }*/
 
-int setTarget(PRGB newTarget, int w, int h) {
-    int last = (int) target;
+long setTarget(PRGB newTarget, int w, int h) {
+    long last = (long) target;
     if (newTarget == 0) {
         target = (PRGB) image;
         width = client_width;
